requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.2
azure-core==1.31.0
azure-storage-blob==12.23.1
bcrypt==4.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
databases==0.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.113.0
fastapi-cli==0.0.5
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
idna==3.8
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
mailjet-rest==1.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
oauthlib==3.2.2
orjson==3.10.7
passlib==1.7.4
pdf2image==1.17.0
pillow==10.4.0
plum-dispatch==1.7.4
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.0
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.2
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyMySQL==1.1.1
pyparsing==3.1.4
pypng==0.20220715.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-pptx==1.0.2
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.4
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
XlsxWriter==3.2.0
